pyproject.toml
[project]
name = "research_analyst_literature_review_generator"
version = "0.1.0"
description = "research_analyst_literature_review_generator using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.70.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "pandas>=2.2.0",
    "requests>=2.31.0",
    "scidownl>=1.0.0",
]

[project.scripts]
research_analyst_literature_review_generator = "research_analyst_literature_review_generator.main:run"
run_crew = "research_analyst_literature_review_generator.main:run"
train = "research_analyst_literature_review_generator.main:train"
replay = "research_analyst_literature_review_generator.main:replay"
test = "research_analyst_literature_review_generator.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
